requirements.txt
fastapi 
uvicorn 
openai 
pydantic 
python-dotenv 
streamlit
requests